Move callbacks closer to Gutenberg fragment
